Updated ImageOriginalXC
ImageOriginalXC was returning a Picture with no Graphics object and zero dimensions. Using older code that had been commented out, which obtained a pointer by using the CopyOSHandle method of the Picture object rather than the Handle property, I changed the way this was working and now I get a Picture with a Graphics object.